refactor(adapter-oas): put discriminator handling behind emit/discriminator/
Discriminator logic lived in three places: the inline union/allOf
narrowing in composition.ts, the buildDiscriminatorChildMap/
patchDiscriminatorNode post-pass in discriminator.ts, and the wiring
between them in adapter.ts. Split by mode instead:

- emit/discriminator/preserve.ts: narrowUnionMembers and
  extractDiscriminatedAllOfMembers, extracted from convertUnion and
  convertAllOf's inline sub-functions. Narrows a union member with its
  discriminant value at the usage site; the referenced child schema's
  own definition is untouched.
- emit/discriminator/propagate.ts: buildDiscriminatorChildMap and
  patchDiscriminatorNode, moved from discriminator.ts unchanged. The
  post-pass adapter.ts runs under discriminator: 'propagate' to push
  the discriminant onto each child schema's own definition.
- emit/discriminator/index.ts: re-exports both plus a DiscriminatorStrategy
  type documenting how the two compose (propagate is preserve's inline
  narrowing plus the post-pass, not a full swap).

convertUnion and convertAllOf now call into preserve.ts instead of
carrying the branching themselves.

Part of the adapter-oas leaner-shape refactor (slice 05/07): no
behavior change, verified by the characterization snapshots.
