+++ packages/adapter-oas/src/adapter.ts
@@ -2,8 +2,8 @@ import { ast, findCircularSchemas, narrowSchema } from '@kubb/ast'
 import { createAdapter } from '@kubb/core'
 import type { AdapterSource } from '@kubb/core'
 import { DEFAULT_PARSER_OPTIONS } from './constants.ts'
-import { buildDiscriminatorChildMap, patchDiscriminatorNode } from './discriminator.ts'
-import type { DiscriminatorTarget } from './discriminator.ts'
+import { buildDiscriminatorChildMap, patchDiscriminatorNode } from './emit/discriminator/index.ts'
+import type { DiscriminatorTarget } from './emit/discriminator/index.ts'
 import { assertInputExists, parseDocument, parseFromConfig, validateDocument } from './factory.ts'
 import { getOperations } from './operation.ts'
 import { createSchemaParser } from './parser.ts'
